requirements.txt
discord.py==2.3.2
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0